// src/components/frame/v5/pages/UserCryptoToFiatPage/partials/Verification/types.ts
// @TODO Might still be subject to updates
export enum KycStatus {
  NOT_STARTED = 'not_started',
  PENDING = 'pending',
  INCOMPLETE = 'incomplete',
  AWAITING_UBO = 'awaiting_ubo',
  MANUAL_REVIEW = 'manual_review',
  UNDER_REVIEW = 'under_review',
  APPROVED = 'approved',
  REJECTED = 'rejected',
}